+++ src/presentation/protocols/controllers.ts
@@ -1,4 +1,4 @@
-import { HttpRequest, HttpResponse } from "../protocols/https";
+import { HttpRequest, HttpResponse } from "./https";
 
 export interface IController {
   handle(httpResponse: HttpRequest): HttpResponse
