+++ src/presentation/protocols/index.ts
@@ -1,0 +1,3 @@
+export * from './https'
+export * from './controllers'
+export * from './email-validation'
